+++ pom.xml
@@ -45,6 +45,7 @@
 
 	<dependencies>
 
+		<!-- Spring Functional Dependencies -->
 		<dependency>
 			<groupId>org.springframework.fu</groupId>
 			<artifactId>spring-fu-jafu</artifactId>
@@ -69,6 +70,9 @@
 			<artifactId>spring-init-library</artifactId>
 		</dependency>
 
+		<!-- /Functional Dependencies -->
+
+		<!-- Boot Dependencies -->
 
 		<dependency>
 			<groupId>org.springframework.security</groupId>
@@ -77,10 +81,6 @@
 		<dependency>
 			<groupId>org.springframework.security</groupId>
 			<artifactId>spring-security-oauth2-jose</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>javax.xml.bind</groupId>
-			<artifactId>jaxb-api</artifactId>
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
@@ -143,11 +143,18 @@
 			<groupId>org.springframework.security</groupId>
 			<artifactId>spring-security-messaging</artifactId>
 		</dependency>
+		
+		<!--/Boot Dependencies -->
 
 		<dependency>
 			<groupId>io.jsonwebtoken</groupId>
 			<artifactId>jjwt</artifactId>
 			<version>0.9.1</version>
+		</dependency>
+
+		<dependency>
+			<groupId>javax.xml.bind</groupId>
+			<artifactId>jaxb-api</artifactId>
 		</dependency>
 	</dependencies>
