+++ src/main/java/com/mangasite/web/routers/TokenRouter.java
@@ -1,4 +1,4 @@
-package com.mangasite.service.jwt;
+package com.mangasite.web.routers;
 
 import static org.springframework.web.reactive.function.server.RequestPredicates.GET;
 import static org.springframework.web.reactive.function.server.RouterFunctions.route;
@@ -6,6 +6,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.reactive.function.server.RouterFunction;
 import org.springframework.web.reactive.function.server.ServerResponse;
+import com.mangasite.services.TokenService;
 
 @Configuration
 public class TokenRouter {
